+++ src/ast/passes/codegen_llvm.cpp
@@ -2724,6 +2724,17 @@ void CodegenLLVM::visit(Probe &probe)
             erase_prefix(match);
             probefull_ = attach_point->name(match);
           }
+          else if (probetype(attach_point->provider) == ProbeType::software ||
+                   probetype(attach_point->provider) == ProbeType::hardware ||
+                   probetype(attach_point->provider) == ProbeType::interval ||
+                   probetype(attach_point->provider) == ProbeType::profile)
+          {
+            // Hardware, software, interval, and profile probes do not support
+            // wildcards but they still may need expansion when the 'probe'
+            // builtin is used.
+            // Just use the name stored in the probe in such a case.
+            probefull_ = attach_point->name("");
+          }
           else
             probefull_ = attach_point->name(match);