+++ src/probe_matcher.cpp
@@ -199,6 +199,14 @@ std::set<std::string> ProbeMatcher::get_matches_for_probetype(
       symbol_stream = std::make_unique<std::istringstream>(ret);
       break;
     }
+    case ProbeType::interval:
+    case ProbeType::profile: {
+      std::string ret;
+      for (auto& unit : TIME_UNITS)
+        ret += unit + ":\n";
+      symbol_stream = std::make_unique<std::istringstream>(ret);
+      break;
+    }
     default:
       return {};
   }
@@ -540,8 +548,6 @@ std::set<std::string> ProbeMatcher::get_matches_for_ap(
     case ProbeType::watchpoint:
     case ProbeType::asyncwatchpoint:
     case ProbeType::tracepoint:
-    case ProbeType::hardware:
-    case ProbeType::software:
     case ProbeType::kfunc:
     case ProbeType::kretfunc: {
       // Do not expand "target:" as that would match all functions in target.
@@ -550,6 +556,13 @@ std::set<std::string> ProbeMatcher::get_matches_for_ap(
         return { attach_point.target + ":" };
 
       search_input = attach_point.target + ":" + attach_point.func;
+      break;
+    }
+    case ProbeType::hardware:
+    case ProbeType::software:
+    case ProbeType::profile:
+    case ProbeType::interval: {
+      search_input = attach_point.target + ":";
       break;
     }
     case ProbeType::usdt:
@@ -576,11 +589,6 @@ std::set<std::string> ProbeMatcher::get_matches_for_ap(
       throw WildcardException(
           "Wildcard matches aren't available on probe type '" +
           attach_point.provider + "'");
-    case ProbeType::profile:
-    case ProbeType::interval:
-      // Wildcard matches are not supported on these probe types, however
-      // expansion can still happen when the probe builtin is used
-      return { "" };
   }
 
   return get_matches_for_probetype(probetype(attach_point.provider),